Renames module to line up with convention
